udf_api/requirements.txt
Flask==3.0.0
openai==1.7.2
python-dotenv==1.0.0